requirements.txt
google-generativeai==0.8.5
edge-tts==7.2.3
gTTS==2.5.4
Pillow==10.4.0
transformers==4.57.6
PyPDF2==3.0.1
python-docx==1.1.2
groq==0.15.0
huggingface-hub==0.36.1
torch==2.5.1
fastapi==0.115.0
uvicorn==0.32.0